pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gh-renovate-silencer"
version = "0.1.0"
description = "A tool to silence GitHub notifications from Renovate bot"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "rwxd", email = "" },
]
dependencies = [
    "typer>=0.9.0",
    "pygithub>=2.1.1",
    "pydantic>=2.4.0",
    "rich>=13.4.2",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
    "pre-commit>=3.5.0",
    "types-requests>=2.31.0",
]

[project.scripts]
gh-renovate-silencer = "gh_renovate_silencer.cli:app"

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "ARG", "SIM", "ERA", "PL", "RUF"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["gh_renovate_silencer"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
